Move EndMessageKnowledge into own file
